pyproject.toml
[project]
name = "gdrive-deduper"
version = "0.1.0"
description = "Find duplicate files in Google Drive using MD5 checksums"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "gradio>=4.0.0",
    "pdf2image>=1.16.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gdrive-deduper = "gdrive_deduper.cli:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
